Rename compile-time-lookup.ts -> resolver-delegate.ts
