+++ packages/@glimmer/application/src/loaders/runtime-compiler/resolver-delegate.ts
